% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiUtils-package.R
\docType{package}
\name{epiUtils-package}
\alias{epiUtils}
\alias{epiUtils-package}
\title{epiUtils: Epidemiological Analysis Utilities}
\description{
A collection of utility functions for epidemiological analysis, including age-standardised rate calculations, person-years computation, and other common epidemiological methods for cohort and population-based studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmgpanw.github.io/epiUtils/}
  \item \url{https://github.com/rmgpanw/epiUtils}
  \item Report bugs at \url{https://github.com/rmgpanw/epiUtils/issues}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick.19@ucl.ac.uk} (\href{https://orcid.org/0000-0002-0800-2890}{ORCID}) [copyright holder]

}
\keyword{internal}
