% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate Input Data}
\usage{
validate(x, ...)
}
\arguments{
\item{x}{Object to validate}

\item{...}{Additional arguments passed to methods}
}
\value{
Returns \code{x} invisibly if validation passes, otherwise throws an error
}
\description{
S3 generic function for validating different types of input data used in
epidemiological calculations. Provides consistent validation with informative
error messages following the tidyverse style guide.
}
